Added code for parsing HTTP POST data into a dict in request.cpp Added a new error code in cgi.hpp Changed some declarations in request.hpp to comply with request.cpp's current state.
The HTTP POST code has to be tested. Also, for CGI::Request::getParam(), in
which order should we process data? GPCSE or something else.

GPCSE = GET POST COOKIE SESSION ENVIRONMENT

The first match will be returned in the order.

Code for COOKIE, SESSION parsing has to be added. If after implementing
cookie, class is big enough, we'll make another class for session.