Session class now implements a singleton design - only one instance at a time.
The compile-error bugs in session have been fixed. The module is still
incomplete because we don't have the db module yet to implement session fully.

Modified the definition of std::map (Dict_t) in cgi.hpp and removed const from
them because it is not possible to modify constant variables.